Pipelined notebook
This is the notebook which develops the complete data pipeline for the Final Project of the course Big Data Tools and Techniques. The work is based on an official public dataset from the Brazilian National Telecommunications Agency (Anatel) and focuses on analyzing how rural public schools with internet access are distributed across Brazilian states, cities, and technologies using a graph database approach.